Se hizo un pequeño refactoring de los TESTs para separarlos por tipo.
* La consulta de RUC tiene un TEST más apropiado
* Se agregó un TEST con MONEDA EXTRANJERA y DESCUENTO
* Se sacó algo que tiene que tiene que estar configurado a nivel de librería y no a nivel de Spring.
